Separate test files by environment
