+++ javascript/src/tests/node/test_cli.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import {cli} from '../cli.js';
+import {cli} from '../../cli.js';
 import {execFile, ExecFileException} from 'child_process';
 import * as path from 'path';
 import stream from 'stream';
-import {loadDefaultParsers} from '../parser.js';
+import {loadDefaultParsers} from '../../parser.js';
 
 type execFileCallBack = {
   error: ExecFileException | null;
